simple clean ups
