+++ tailcall-tracker/src/collect/mod.rs
@@ -1,4 +1,3 @@
 pub mod collectors;
 mod ga;
-mod ga_event;
 mod posthog;
