+++ tailcall-tracker/src/error.rs
@@ -1,5 +1,3 @@
-use std::fmt::Display;
-
 use derive_more::{DebugCustom, From};
 use reqwest::header::InvalidHeaderValue;
 
@@ -22,19 +20,6 @@ pub enum Error {
 
     #[debug(fmt = "Tokio Join Error: {}", _0)]
     TokioJoin(tokio::task::JoinError),
-}
-
-impl Display for Error {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Error::Reqwest(error) => write!(f, "Reqwest Error: {}", error),
-            Error::InvalidHeaderValue(error) => write!(f, "Invalid Header Value: {}", error),
-            Error::SerdeJson(error) => write!(f, "Serde JSON Error: {}", error),
-            Error::UrlParser(error) => write!(f, "Url Parser Error: {}", error),
-            Error::PostHog(error) => write!(f, "PostHog Error: {}", error),
-            Error::TokioJoin(error) => write!(f, "Tokio Join Error: {}", error),
-        }
-    }
 }
 
 pub type Result<A> = std::result::Result<A, Error>;
